feat(image-gen, database): Upgrade Stable Diffusion XL & refactor furniture schema
This commit includes:

- Upgraded AI image generation model from Stable Diffusion v1.5 to Stable Diffusion XL Refiner 1.0.

- Adjusted generated image resolution from (512, 512) to (1280, 720).

- Renamed 'furniture' table to 'furniture_database' in models/schemas for clarity.

- Added a 'type' column to the 'furniture_coordinates' table.

- Cleaned up old AI trial files (generated_v1.py, generated_v2.py) and removed yolo11n.pt.
