+++ backend/schemas/schemas.py
@@ -6,7 +6,8 @@ from datetime import datetime
 
 # Pydantic Schema: Represents the structure of a furniture item
 
-class FurnitureCreate(BaseModel):
+# --- Furniture Database (Renamed) --- 
+class FurnitureDatabaseCreate(BaseModel):
   furniture_id: Optional[str] = None
   style: Optional[str] = None
   room: Optional[str] = None
@@ -19,7 +20,7 @@ class FurnitureCreate(BaseModel):
   class Config:
     from_attributes = True # Allows Pydantic to work with SQLAlchemy models directly
 
-class FurnitureModel(FurnitureCreate):
+class FurnitureDatabaseModel(FurnitureDatabaseCreate):
   id: int
 
 # --- Pydantic Schemas for Generated Rooms ---
@@ -47,6 +48,7 @@ class FurnitureCoordinatesModel(BaseModel):
   furniture_id: str
   x_coordinate: float
   y_coordinate: float
+  type: Optional[str] = None
 
   class Config:
     from_attributes = True # Allows Pydantic to work with SQLAlchemy models directly
@@ -55,6 +57,7 @@ class FurnitureCoordinateCreate(BaseModel):
   furniture_id: str
   x_coordinate: float
   y_coordinate: float
+  type: Optional[str] = None
 
 class FurnitureCoordinateBatchCreate(BaseModel):
   generated_room_id: str
